others/abc084/d/src/main/java/com/solver/atcoder/others/abc084/d/Main.java
package com.solver.atcoder.others.abc084.d;

import java.io.InputStream;
import java.io.IOException;
import java.io.PrintStream;
import java.util.*;

public class Main {

  public static void main(String[] args) {
    solve(System.in, System.out);
  }

  private static final int n = 100000;

  static void solve(InputStream is, PrintStream os) {
    Scanner sc = new Scanner(is);

    /* read */
    Set<Integer> primes = new HashSet<>(primes());
    int[] culm = new int[n + 1];
    // int occur = 0;
    for (int i = 1; i <= n; i += 2) {
      if (primes.contains(i) && primes.contains((i + 1) / 2)) culm[i]++;
    }

    for (int i = 1; i <= n; i++) {
      culm[i] += culm[i - 1];
    }

    int q = sc.nextInt();
    for (int i = 0; i < q; i++) {
      int l = sc.nextInt();
      int r = sc.nextInt();
      os.println(culm[r] - culm[l - 1]);
    }
  }

  private static List<Integer> primes() {
    List<Integer> primes = new ArrayList<>();
    List<Integer> ints = new LinkedList<>();

    for (int i = 2; i <= n; i++) {
      ints.add(i);
    }

    for (int i = 2; i * i <= n; i++) {
      int prime = ints.remove(0);
      primes.add(prime);
      List<Integer> nexts = new LinkedList<>();
      for (int next : ints) {
        if (next % prime != 0) nexts.add(next);
      }
      ints = nexts;
    }
    primes.addAll(ints);
    return primes;
  }

  private static class Scanner {

    private final InputStream is;
    private final byte[] buffer = new byte[1024];
    private int ptr = 0;
    private int buflen = 0;

    Scanner(InputStream is) {
      this.is = is;
    }

    private boolean hasNextByte() {
      if (ptr < buflen) {
        return true;
      } else {
        ptr = 0;
        try {
          buflen = is.read(buffer);
        } catch (IOException e) {
          e.printStackTrace();
        }
        if (buflen <= 0) {
          return false;
        }
      }
      return true;
    }

    private int readByte() {
      if (hasNextByte()) {
        return buffer[ptr++];
      } else {
        return -1;
      }
    }

    private static boolean isPrintableChar(int c) {
      return 33 <= c && c <= 126;
    }

    public boolean hasNext() {
      while (hasNextByte() && !isPrintableChar(buffer[ptr])) {
        ptr++;
      }
      return hasNextByte();
    }

    public String next() {
      if (!hasNext()) {
        throw new NoSuchElementException();
      }
      StringBuilder sb = new StringBuilder();
      int b = readByte();
      while (isPrintableChar(b)) {
        sb.appendCodePoint(b);
        b = readByte();
      }
      return sb.toString();
    }

    public long nextLong() {
      if (!hasNext()) {
        throw new NoSuchElementException();
      }
      long n = 0;
      boolean minus = false;
      int b = readByte();
      if (b == '-') {
        minus = true;
        b = readByte();
      }
      if (b < '0' || '9' < b) {
        throw new NumberFormatException();
      }
      while (true) {
        if ('0' <= b && b <= '9') {
          n *= 10;
          n += b - '0';
        } else if (b == -1 || !isPrintableChar(b)) {
          return minus ? -n : n;
        } else {
          throw new NumberFormatException();
        }
        b = readByte();
      }
    }

    public int nextInt() {
      long nl = nextLong();
      if (nl < Integer.MIN_VALUE || nl > Integer.MAX_VALUE) {
        throw new NumberFormatException();
      }
      return (int) nl;
    }

    public double nextDouble() {
      return Double.parseDouble(next());
    }
  }
}
